Consolidate duplicated process-spawning test helpers into tests/support
aver_bin, repo_root, format_output, cleanup, temp_module, and aver_command
were redeclared, byte-identical (or identical modulo fs:: vs std::fs:: and
a variable name), across 38 integration suites. Move the single copy into
tests/support/aver_cmd.rs, included per suite via #[path], and delete the
local copies; #![allow(dead_code)] on the module accounts for each suite
using only the subset of helpers it needs.

Left every run/run_vm/run_aver/run_wasm_gc variant and every temp_dir or
temp_module body that differs in naming or sets suite-specific env
(AVER_NO_CARRIER_I64, CARGO_NET_OFFLINE, AVER_CERT_PRELUDE_CACHE, etc.)
untouched, and did not merge any test binaries — CI names several wasm-gc
suites individually.

Verified with cargo check --features wasm,wasip2 (clean), cargo clippy
--features wasm,wasip2 (identical warning count and locations to a
git-stash baseline), and full runs of parser_spec, vm_spec,
agent_connect_spec, and compile_spec.
