+++ tests/cert_certify_spec.rs
@@ -9,6 +9,11 @@
 //! backend) and skipped when `lake` is unavailable, mirroring `proof_spec.rs`.
 #![cfg(feature = "wasm")]
 
+#[path = "support/aver_cmd.rs"]
+mod aver_cmd;
+
+use aver_cmd::aver_command;
+
 #[path = "support/cert_wall.rs"]
 mod cert_wall;
 #[path = "support/scratch_dir.rs"]
@@ -19,19 +24,6 @@ use scratch_dir::{ScratchDir, temp_dir};
 use std::collections::{BTreeMap, BTreeSet};
 use std::path::PathBuf;
 use std::process::Command;
-
-fn aver_command() -> Command {
-    let mut command = Command::new(env!("CARGO_BIN_EXE_aver"));
-    command.env(
-        "AVER_CERT_PRELUDE_CACHE",
-        std::env::temp_dir().join("aver-cert-prelude-store"),
-    );
-    command.env(
-        "AVER_CERT_DATA_CACHE",
-        std::env::temp_dir().join("aver-cert-data-store"),
-    );
-    command
-}
 
 fn copy_dir_all(src: &std::path::Path, dst: &std::path::Path) {
     std::fs::create_dir_all(dst).unwrap();
