+++ tests/cert_whole_module_guard_iso.rs
@@ -1,5 +1,10 @@
 #![cfg(feature = "wasm")]
 
+#[path = "support/aver_cmd.rs"]
+mod aver_cmd;
+
+use aver_cmd::aver_command;
+
 #[path = "support/cert_wall.rs"]
 mod cert_wall;
 #[path = "support/scratch_dir.rs"]
@@ -10,19 +15,6 @@ use scratch_dir::temp_dir;
 use std::collections::BTreeSet;
 use std::path::{Path, PathBuf};
 use std::process::Command;
-
-fn aver_command() -> Command {
-    let mut command = Command::new(env!("CARGO_BIN_EXE_aver"));
-    command.env(
-        "AVER_CERT_PRELUDE_CACHE",
-        std::env::temp_dir().join("aver-cert-prelude-store"),
-    );
-    command.env(
-        "AVER_CERT_DATA_CACHE",
-        std::env::temp_dir().join("aver-cert-data-store"),
-    );
-    command
-}
 
 fn assert_manifest_decode_declines(wasm: &std::path::Path, cert: &std::path::Path, expected: &str) {
     let output = aver_command()
