+++ tests/cross_backend_proptest.rs
@@ -26,10 +26,14 @@
 //!   gate fires there.
 #![cfg(feature = "wasm")]
 
+#[path = "support/aver_cmd.rs"]
+mod aver_cmd;
+
+use aver_cmd::{format_output, cleanup, temp_module};
+
 use std::fs;
 use std::path::PathBuf;
 use std::process::Command;
-use std::time::{SystemTime, UNIX_EPOCH};
 
 use proptest::prelude::*;
 
@@ -40,31 +44,6 @@ use proptest::prelude::*;
 // crate, sharing requires a `mod common;` ref from every test file plus
 // the visibility plumbing. The runners are small enough that the
 // duplication does not pay back the refactor.
-
-fn temp_module(prefix: &str, source: &str) -> PathBuf {
-    let nanos = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .expect("system time before unix epoch")
-        .as_nanos();
-    let dir = std::env::temp_dir().join(format!("{}-{}", prefix, nanos));
-    fs::create_dir_all(&dir).expect("create temp dir");
-    let path = dir.join("main.av");
-    fs::write(&path, source).expect("write temp module source");
-    path
-}
-
-fn cleanup(path: &std::path::Path) {
-    let _ = fs::remove_dir_all(path.parent().expect("temp module has parent"));
-}
-
-fn format_output(out: &std::process::Output) -> String {
-    format!(
-        "status: {}\nstdout:\n{}\nstderr:\n{}",
-        out.status,
-        String::from_utf8_lossy(&out.stdout),
-        String::from_utf8_lossy(&out.stderr),
-    )
-}
 
 fn run_vm(prefix: &str, source: &str) -> Result<String, String> {
     let repo_root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
