+++ tests/rust_codegen_differential.rs
@@ -61,6 +61,11 @@
 
 #![cfg(feature = "runtime")]
 
+#[path = "support/aver_cmd.rs"]
+mod aver_cmd;
+
+use aver_cmd::{aver_bin, repo_root, format_output};
+
 use std::fs;
 use std::path::{Path, PathBuf};
 use std::process::Command;
@@ -73,14 +78,6 @@ use std::time::{SystemTime, UNIX_EPOCH};
 /// temp-dir name even within the same nanosecond.
 static UNIQUE: AtomicU64 = AtomicU64::new(0);
 
-fn repo_root() -> PathBuf {
-    PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-}
-
-fn aver_bin() -> &'static str {
-    env!("CARGO_BIN_EXE_aver")
-}
-
 fn temp_dir(prefix: &str) -> PathBuf {
     let nanos = SystemTime::now()
         .duration_since(UNIX_EPOCH)
@@ -90,15 +87,6 @@ fn temp_dir(prefix: &str) -> PathBuf {
     let dir = std::env::temp_dir().join(format!("aver-rust-diff-{prefix}-{nanos}-{n}"));
     fs::create_dir_all(&dir).expect("create temp dir");
     dir
-}
-
-fn format_output(output: &std::process::Output) -> String {
-    format!(
-        "status: {}\nstdout:\n{}\nstderr:\n{}",
-        output.status,
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    )
 }
 
 /// Single `cargo build` target dir shared across every example in one
