+++ tests/verify_unchecked_files_spec.rs
@@ -6,12 +6,13 @@
 //! The negative control at the bottom is load-bearing: a clean run must print
 //! exactly what it printed before, byte for byte.
 
+#[path = "support/aver_cmd.rs"]
+mod aver_cmd;
+
+use aver_cmd::aver_bin;
+
 use std::fs;
 use std::process::Command;
-
-fn aver_bin() -> &'static str {
-    env!("CARGO_BIN_EXE_aver")
-}
 
 fn temp_dir(tag: &str) -> std::path::PathBuf {
     let dir = std::env::temp_dir().join(format!(
